requirements.txt
Flask==2.0.2
Flask-MySQLdb==0.2.0
captcha==0.3
requests==2.26.0
mysql-connector-python==8.0.27
